[#99] refactor: 이름 변경 및 람다로 변경 - OAuth2Request -> LoginOAuth2Request 이름 변경 - 람다 사용, 간단 리팩토링
