+++ src/main/java/goorm/eagle7/stelligence/common/login/LoginService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import goorm.eagle7.stelligence.common.auth.jwt.JwtTokenProvider;
+import goorm.eagle7.stelligence.common.login.dto.LoginOAuth2Request;
 import goorm.eagle7.stelligence.common.login.dto.LoginTokensWithIdAndRoleResponse;
 import goorm.eagle7.stelligence.common.util.CookieType;
 import goorm.eagle7.stelligence.common.util.CookieUtils;
@@ -25,25 +26,25 @@ public class LoginService {
 	private final CookieUtils cookieutils;
 
 	@Transactional
-	public Member login(HttpServletResponse response, OAuth2Request oAuth2Request) {
+	public Member login(HttpServletResponse response, LoginOAuth2Request loginOAuth2Request) {
 
-		String socialId = oAuth2Request.getSocialId();
+		String socialId = loginOAuth2Request.getSocialId();
 
 		// TODO socialId, socialType으로 회원 조회 후 없으면 회원 가입 -> member 받아 오기
 		// socialId로 회원 조회 후 없으면 회원 가입 -> member 받아 오기
 		Member member = memberRepository.findBySocialId(socialId)
-			.orElseGet(() -> signUpService.signUp(oAuth2Request));
+			.orElseGet(() -> signUpService.signUp(loginOAuth2Request));
 
 		// token 생성 후 refreshToken DB에 저장
 		LoginTokensWithIdAndRoleResponse loginTokensWithIdAndRoleResponse = generateAndSaveTokens(member);
 
 		// 발행된 토큰을 사용자의 브라우저 쿠키에 저장
-		addTokensOnCookies(response, loginTokensWithIdAndRoleResponse);
+		addTokensOnCookies(loginTokensWithIdAndRoleResponse);
 
 		return member;
 	}
 
-	private void addTokensOnCookies(HttpServletResponse response, LoginTokensWithIdAndRoleResponse loginTokensWithIdAndRoleResponse) {
+	private void addTokensOnCookies(LoginTokensWithIdAndRoleResponse loginTokensWithIdAndRoleResponse) {
 
 		// 토큰 추출
 		String accessToken = loginTokensWithIdAndRoleResponse.getAccessToken();
@@ -59,7 +60,6 @@ public class LoginService {
 	 * @param member 회원
 	 * @return accessToken, refreshToken
 	 */
-	@Transactional
 	protected LoginTokensWithIdAndRoleResponse generateAndSaveTokens(Member member) {
 
 		// Token 생성
@@ -83,7 +83,7 @@ public class LoginService {
 
 		// db에서 refreshToken 삭제
 		memberRepository.findById(memberId)
-			.ifPresent(member -> member.expireRefreshToken());
+			.ifPresent(Member::expireRefreshToken);
 
 	}
